Move and refactor shapes from sensormanagement to types
